requirements.txt
# Core
Django==5.0.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9

# Authentification
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Stockage & Médias
django-storages==1.13.2
boto3==1.28.74

# Base de données
dj-database-url==2.1.0

# Utilitaires
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Développement (optionnel)
python-dotenv==1.0.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0